+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
-## [Unreleased]
+## [0.6.1] - 2024-12-13
 
 ### Added
 - [docs] Add documentation link for docs.rs in package metadata
@@ -135,7 +135,8 @@ Initial release of pleezer, a headless streaming player for the Deezer Connect p
 - Debug and trace logging capabilities
 - Command-line interface with various configuration options
 
-[Unreleased]: https://github.com/roderickvd/pleezer/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/roderickvd/pleezer/compare/v0.6.1...HEAD
+[0.6.0]: https://github.com/roderickvd/pleezer/releases/tag/v0.6.1
 [0.6.0]: https://github.com/roderickvd/pleezer/releases/tag/v0.6.0
 [0.5.0]: https://github.com/roderickvd/pleezer/releases/tag/v0.5.0
 [0.4.0]: https://github.com/roderickvd/pleezer/releases/tag/v0.4.0